extract gql_tristate_value to separate package in order to not pollute gql_exec for users who don't use tristate null (#426)
* extract gql_tristate_value to separate package in order to not pollute gql_exec for users who don't use tristate null

* add lint dependency
