+++ codegen/end_to_end_test_tristate/pubspec.yaml
@@ -15,3 +15,4 @@ dev_dependencies:
   build: ^2.0.0
   build_runner: ^2.0.0
   test: ^1.16.8
+  gql_tristate_value: ^1.0.0
