+++ codegen/end_to_end_test_tristate/test/schema/scalars_test.dart
@@ -1,5 +1,5 @@
 import 'package:built_collection/built_collection.dart';
-import 'package:gql_exec/value.dart';
+import 'package:gql_tristate_value/gql_tristate_value.dart';
 import "package:test/test.dart";
 
 import 'package:end_to_end_test_tristate/graphql/__generated__/schema.schema.gql.dart';
