+++ codegen/gql_build/pubspec.yaml
@@ -17,6 +17,7 @@ dependencies:
   gql_code_builder: ^0.8.0
   path: ^1.8.0
   yaml: ^3.1.0
+  gql_tristate_value: ^1.0.0
 dev_dependencies: 
   build_test: ^2.0.0
   gql_pedantic: ^1.0.2
