+++ codegen/gql_code_builder/pubspec.yaml
@@ -13,6 +13,7 @@ dependencies:
   gql: ^1.0.0
   gql_exec: ^1.0.0
   path: ^1.8.0
+  gql_tristate_value: ^1.0.0
 dev_dependencies: 
   build_runner: ^2.1.0
   gql_pedantic: ^1.0.2
