+++ codegen/gql_tristate_value/analysis_options.yaml
@@ -1,0 +1,1 @@
+include: package:lints/recommended.yaml
